requirements.txt
langchain
langchain-groq==0.2.5
python-dotenv==1.0.0
groq==0.19.0
pydantic==2.10.6
httpx==0.28.1
flask==3.0.0
flask-cors==4.0.0 